Refactor common feed types
